@@ -1,23 +1,23 @@
-import './App.css';
-import Sidebar from './components/Sidebar';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import Overview from './pages/Overview';
-import { Reports, ReportsOne, ReportsTwo, ReportsThree } from './pages/Reports';
-import Team from './pages/Team';
+import { Route, Routes } from "react-router-dom";
+import "./App.css";
+import Sidebar from "./components/Sidebar";
+import Overview from "./pages/Overview";
+import { Reports, ReportsOne, ReportsThree, ReportsTwo } from "./pages/Reports";
+import Team from "./pages/Team";
 
 function App() {
   return (
-    <Router>
+    <div>
       <Sidebar />
-      <Switch>
-        <Route path='/overview' exact component={Overview} />
-        <Route path='/reports' exact component={Reports} />
-        <Route path='/reports/reports1' exact component={ReportsOne} />
-        <Route path='/reports/reports2' exact component={ReportsTwo} />
-        <Route path='/reports/reports3' exact component={ReportsThree} />
-        <Route path='/team' exact component={Team} />
-      </Switch>
-    </Router>
+      <Routes>
+        <Route path="/overview" element={<Overview />} />
+        <Route path="/reports" element={<Reports />} />
+        <Route path="/reports/reports1" element={<ReportsOne />} />
+        <Route path="/reports/reports2" element={<ReportsTwo />} />
+        <Route path="/reports/reports3" element={<ReportsThree />} />
+        <Route path="/team" element={<Team />} />
+      </Routes>
+    </div>
   );
 }
 
